bug fixed for BeanToArray & refactor testcase
