+++ src/test/java/com/alibaba/json/bvt/parser/deser/awt/FontDeserializerTest.java
@@ -1,7 +1,6 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.awt;
 
 import java.awt.Font;
-import java.awt.Rectangle;
 
 import org.junit.Assert;
 
@@ -12,20 +11,20 @@ import com.alibaba.fastjson.serializer.AwtCodec;
 import junit.framework.TestCase;
 
 
-public class RectangleDeserializerTest extends TestCase {
+public class FontDeserializerTest extends TestCase {
     public void test_0 () throws Exception {
         AwtCodec.instance.getFastMatchToken();
         
-        Assert.assertNull(JSON.parseObject("null", Rectangle.class));
-        Assert.assertNull(JSON.parseArray("null", Rectangle.class));
-        Assert.assertNull(JSON.parseArray("[null]", Rectangle.class).get(0));
+        Assert.assertNull(JSON.parseObject("null", StackTraceElement.class));
+        Assert.assertNull(JSON.parseArray("null", StackTraceElement.class));
+        Assert.assertNull(JSON.parseArray("[null]", StackTraceElement.class).get(0));
         Assert.assertNull(JSON.parseObject("{\"value\":null}", VO.class).getValue());
     }
     
     public void test_stack_error_0() throws Exception {
         Exception error = null;
         try {
-            JSON.parseObject("[]", Rectangle.class);
+            JSON.parseObject("[]", Font.class);
         } catch (JSONException ex) {
             error = ex;
         }
@@ -35,7 +34,7 @@ public class RectangleDeserializerTest extends TestCase {
     public void test_stack_error_1() throws Exception {
         Exception error = null;
         try {
-            JSON.parseObject("{33:22}", Rectangle.class);
+            JSON.parseObject("{33:22}", Font.class);
         } catch (JSONException ex) {
             error = ex;
         }
@@ -45,7 +44,7 @@ public class RectangleDeserializerTest extends TestCase {
     public void test_stack_error_2() throws Exception {
         Exception error = null;
         try {
-            JSON.parseObject("{\"name\":22}", Rectangle.class);
+            JSON.parseObject("{\"name\":22}", Font.class);
         } catch (JSONException ex) {
             error = ex;
         }
@@ -55,7 +54,7 @@ public class RectangleDeserializerTest extends TestCase {
     public void test_stack_error_3() throws Exception {
         Exception error = null;
         try {
-            JSON.parseObject("{\"style\":true}", Rectangle.class);
+            JSON.parseObject("{\"style\":true}", Font.class);
         } catch (JSONException ex) {
             error = ex;
         }
@@ -65,7 +64,7 @@ public class RectangleDeserializerTest extends TestCase {
     public void test_stack_error_4() throws Exception {
         Exception error = null;
         try {
-            JSON.parseObject("{\"size\":\"33\"}", Rectangle.class);
+            JSON.parseObject("{\"size\":\"33\"}", Font.class);
         } catch (JSONException ex) {
             error = ex;
         }
@@ -84,13 +83,13 @@ public class RectangleDeserializerTest extends TestCase {
     
     public static class VO {
 
-        private Rectangle value;
+        private Font value;
 
-        public Rectangle getValue() {
+        public Font getValue() {
             return value;
         }
 
-        public void setValue(Rectangle value) {
+        public void setValue(Font value) {
             this.value = value;
         }
