+++ src/test/java/com/alibaba/json/bvt/parser/deser/awt/PointDeserializerTest2.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.awt;
 
 import java.awt.Point;
