+++ src/test/java/com/alibaba/json/bvt/parser/deser/date/DateParseTest6.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.date;
 
 import java.util.Calendar;
 import java.util.Date;
@@ -11,14 +11,14 @@ import junit.framework.TestCase;
 import com.alibaba.fastjson.JSON;
 
 
-public class DateParseTest4 extends TestCase {
+public class DateParseTest6 extends TestCase {
     protected void setUp() throws Exception {
         JSON.defaultTimeZone = TimeZone.getTimeZone("Asia/Shanghai");
         JSON.defaultLocale = Locale.CHINA;
     }
     
     public void test_date() throws Exception {
-        String text = "\"1979-07-14T13:07:23\"";
+        String text = "\"19790714\"";
         Date date = JSON.parseObject(text, Date.class);
         Calendar calendar = Calendar.getInstance(JSON.defaultTimeZone, JSON.defaultLocale);
         calendar.setTime(date);
@@ -27,9 +27,9 @@ public class DateParseTest4 extends TestCase {
         Assert.assertEquals(6, calendar.get(Calendar.MONTH));
         Assert.assertEquals(14, calendar.get(Calendar.DAY_OF_MONTH));
         
-        Assert.assertEquals(13, calendar.get(Calendar.HOUR_OF_DAY));
-        Assert.assertEquals(7, calendar.get(Calendar.MINUTE));
-        Assert.assertEquals(23, calendar.get(Calendar.SECOND));
+        Assert.assertEquals(0, calendar.get(Calendar.HOUR_OF_DAY));
+        Assert.assertEquals(0, calendar.get(Calendar.MINUTE));
+        Assert.assertEquals(0, calendar.get(Calendar.SECOND));
         Assert.assertEquals(0, calendar.get(Calendar.MILLISECOND));
     }
 }
