+++ src/test/java/com/alibaba/json/bvt/parser/deser/list/ArrayDeserializerTest.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.list;
 
 import org.junit.Assert;
 import junit.framework.TestCase;
